porting some tests over
