+++ airflow-core/tests/unit/plugins/test_plugins_manager.py
@@ -266,38 +266,6 @@ class TestPluginsManager:
 
         assert caplog.record_tuples == []
 
-    def test_entrypoint_plugin_errors_dont_raise_exceptions(self, mock_metadata_distribution, caplog):
-        """
-        Test that Airflow does not raise an error if there is any Exception because of a plugin.
-        """
-        from airflow.plugins_manager import _load_entrypoint_plugins
-
-        mock_dist = mock.Mock()
-        mock_dist.metadata = {"Name": "test-dist"}
-
-        mock_entrypoint = mock.Mock()
-        mock_entrypoint.name = "test-entrypoint"
-        mock_entrypoint.group = "airflow.plugins"
-        mock_entrypoint.module = "test.plugins.test_plugins_manager"
-        mock_entrypoint.load.side_effect = ImportError("my_fake_module not found")
-        mock_dist.entry_points = [mock_entrypoint]
-
-        with (
-            mock_metadata_distribution(return_value=[mock_dist]),
-            caplog.at_level(logging.ERROR, logger="airflow.plugins_manager"),
-        ):
-            _, import_errors = _load_entrypoint_plugins()
-
-            received_logs = caplog.text
-            # Assert Traceback is shown too
-            assert "Traceback (most recent call last):" in received_logs
-            assert "my_fake_module not found" in received_logs
-            assert "Failed to import plugin test-entrypoint" in received_logs
-            assert (
-                "test.plugins.test_plugins_manager",
-                "my_fake_module not found",
-            ) in import_errors.items()
-
     def test_registering_plugin_macros(self, request):
         """
         Tests whether macros that originate from plugins are being registered correctly.
@@ -381,35 +349,3 @@ class TestPluginsManager:
         with mock.patch("airflow.plugins_manager._load_plugins_from_plugin_directory", return_value=([], [])):
             plugins = plugins_manager._get_plugins()[0]
         assert len(plugins) == 4
-
-
-class TestPluginsDirectorySource:
-    def test_should_return_correct_path_name(self):
-        from airflow import plugins_manager
-
-        source = plugins_manager.PluginsDirectorySource(__file__)
-        assert source.path == "test_plugins_manager.py"
-        assert str(source) == "$PLUGINS_FOLDER/test_plugins_manager.py"
-        assert source.__html__() == "<em>$PLUGINS_FOLDER/</em>test_plugins_manager.py"
-
-
-class TestEntryPointSource:
-    def test_should_return_correct_source_details(self, mock_metadata_distribution):
-        from airflow import plugins_manager
-
-        mock_entrypoint = mock.Mock()
-        mock_entrypoint.name = "test-entrypoint-plugin"
-        mock_entrypoint.module = "module_name_plugin"
-
-        mock_dist = mock.Mock()
-        mock_dist.metadata = {"Name": "test-entrypoint-plugin"}
-        mock_dist.version = "1.0.0"
-        mock_dist.entry_points = [mock_entrypoint]
-
-        with mock_metadata_distribution(return_value=[mock_dist]):
-            plugins_manager._load_entrypoint_plugins()
-
-        source = plugins_manager.EntryPointSource(mock_entrypoint, mock_dist)
-        assert str(mock_entrypoint) == source.entrypoint
-        assert "test-entrypoint-plugin==1.0.0: " + str(mock_entrypoint) == str(source)
-        assert "<em>test-entrypoint-plugin==1.0.0:</em> " + str(mock_entrypoint) == source.__html__()
